Added removeDecimal() to Weather
