+++ app.js
@@ -1,9 +1,13 @@
-//const {Weather} = require('./weather');
+import { Weather } from './weather.js';
 
-
+// GLOBAL VARIABLES
 const APIkey = 'f18b6ae1c57f039f48f95ade89757557';
 
-const weatherObject = {}
+const form = document.getElementById('form');
+
+let unit = undefined;
+
+const weatherObject = {};
 
 const colorScheme = [
     { name: 'crimson'       , rgb: 'rgb(220,20,60)'  , hex: '#DC143C' },
@@ -12,78 +16,53 @@ const colorScheme = [
     { name: 'seagreen'      , rgb: 'rgb(46,139,87)'  , hex: '#2E8B57' },
     { name: 'skyblue'       , rgb: 'rgb(135,206,235)', hex: '#87CEEB' },
     { name: 'indigo'        , rgb: 'rgb(75,0,130)'   , hex: '#4B0082' }
-]
+];
 
 
+//Functions
 
-const form = document.getElementById('form');
-let unit = undefined;
-
-class Weather {
-
-    constructor(weatherObject) {
-        this.weather = weatherObject;
-    }
-
-    displayWeather() {
-        const {cityName, currentTemp, maxTemp, minTemp, feelsLike, humidity, windspeed, description, icon} = this.weather;
-    
-        const iconURL = "http://openweathermap.org/img/w/" + icon + ".png";
-        const newDiv = document.createElement("div");
-        newDiv.class = 'weather-card';
-        let ifTrue = true;
-        while (ifTrue) {
-            newDiv.innerHTML = 
-            `
-            <div class='card'>
-                <h2 class='header'>${cityName}</h2>
-                <img id='icon' src=${iconURL} alt='weather icon'>
-                <p class='card-current-temp'><strong>Current Temp:</strong> ${currentTemp}</p>
-                <p class='card-description'><strong>Description:</strong> ${description}</p>
-                <p class='card-max-temp'><strong>Max Temp:</strong> ${maxTemp}</p>
-                <p class='card-min-temp'><strong>Min Temp:</strong> ${minTemp}</p>
-                <p class='card-feels-like'><strong>Feels Like:</strong> ${feelsLike}</p>
-                <p class='card-humidity'><strong>Humidity:</strong> ${humidity}</p>
-                <p class='card-windspeed'><strong>Windspeed:</strong> ${windspeed}</p>
-            </div`
-        
-            const section = document.getElementById('weather-container');
-            section.appendChild(newDiv);
-
-            ifTrue = false
-        }
-
-    
-    }
-}
-
-
+// @descriptions: 
+// @params: none
+// @return: API data in json format
 const getData = async () => {
     try {
         const cityName = inputValue();
-        const unit = 'metric' //unit
+        const unit = 'metric';
         const response = await fetch(`https://api.openweathermap.org/data/2.5/weather?q=${cityName}&units=${unit}&appid=${APIkey}`);
         const data = await response.json();
-        return data
+        return data;
     } catch (err) {
-        throw new Error(console.error(err.message))
+        throw new Error(console.error(err.message));
     }
 }
 
+// @descriptions: 
+// @params: none
+// @return: return user input value of city name
 const inputValue = () => {
     const cityName = document.getElementById('input').value;
-    return cityName
+    return cityName;
 }
 
+// @descriptions: Console.log the error that may occur when getting API
+// @params: error
+// @return: 
 const apiError = (error) => {
     console.error(error);
+    return;
 }
 
+// @descriptions: Console.log the JSON data from API
+// @params: data
+// @return
 const displayData = (data) => {
-    console.log(`API DATA: ${data}`)
-    return
+    console.log(`API DATA: ${data}`);
+    return;
 }
 
+// @descriptions: 
+// @params:
+// @return
 const formHandler = (event) => {
     event.preventDefault();
     getData().then(displayData).catch(apiError);
@@ -91,14 +70,17 @@ const formHandler = (event) => {
 
 }
 
+// @descriptions: 
+// @params: event
+// @return:
 const formHandler2 = (event) => {
     event.preventDefault();
     getData().then(getWeather).catch(apiError);
 }
 
-form.addEventListener('submit', formHandler)
-
-
+// @descriptions: Assembles data from JSON/API into an object literal to easier make weather card
+// @params: getData function which returns JSON data
+// @return: weatherObject Object Literal
 const asssembleData = async (getData) => {
     const {main,wind,weather,name} = await getData();
     weatherObject.cityName    = name;
@@ -111,13 +93,32 @@ const asssembleData = async (getData) => {
     weatherObject.description = weather[0].description;
     weatherObject.icon        = weather[0].icon;
 
-    return 
+    return;
 }
 
+// @descriptions: 
+// @params: none
+// @return
 const createCard = () => {
     console.log(weatherObject.currentTemp)
-    //const weather = new Weather()
+
 }
+
+// @descriptions: Create Card using Weather Class
+// @params: none
+// @return: 
+const createWeatherObject =  async () => {
+    
+    const weather =  new Weather(weatherObject)
+
+    weather.displayWeather();
+
+    return;
+}
+
+// Event Listener
+form.addEventListener('submit', formHandler)
+
 
 document.getElementById("metric-button").onclick = function() {
     //  unit = "metric"
@@ -126,14 +127,3 @@ document.getElementById("metric-button").onclick = function() {
 document.getElementById("imperial-button").onclick = function() {
     // show imperial
 };
-
-const createWeatherObject =  async () => {
-    
-    const weather =  new Weather(weatherObject)
-
-    weather.displayWeather();
-
-}
-
-
-form.addEventListener('submit', formHandler );
